refactor: SelectionArea.tsx を SRP に基づいて分割 (#205)
- SlotCard を src/components/selection/SlotCard.tsx に分離
- ProgressDots を src/components/selection/ProgressDots.tsx に分離
- MobileCollapsedBar を src/components/selection/MobileCollapsedBar.tsx に分離
- 折りたたみ状態ロジックを useMobileCollapse hook に抽出
- SelectionArea を selection/ 配下のオーケストレーターとして再構成
- CATEGORY_ICONS, CATEGORIES を constants/category.ts に移動
- 各コンポーネント・hookのユニットテストを追加 (16テスト増)
